hack_etherscan/requirements.txt
alabaster==0.7.10
alembic==0.8.9
amqp==1.4.9
anaconda-client==1.6.2
anaconda-navigator==1.5.2
anaconda-project==0.4.1
anyjson==0.3.3
appnope==0.1.0
appscript==1.0.1
apptools==4.4.0
argh==0.26.2
astroid==1.4.9
astropy==1.3.1
attrs==17.2.0
Authomatic==0.1.0.post1
autobahn==17.6.2
Automat==0.6.0
awscli==1.11.125
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.2
beautifulsoup4==4.5.3
billiard==3.3.0.23
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
blinker==1.3
bokeh==0.12.4
boto==2.46.1
botocore==1.5.88
Bottleneck==1.2.0
celery==3.1.18
certifi==2017.7.27.1
cffi==1.9.1
chardet==3.0.4
chest==0.2.3
click==6.6
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.7
conda==4.3.21
configobj==5.0.6
constantly==15.1.0
contextlib2==0.5.4
coverage==4.2
cryptography==1.7.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.1
datashape==0.5.4
decorator==4.0.11
dill==0.2.5
Django==1.11.4
django-async==0.7.3
django-bootstrap3==8.2.3
django-ckeditor==5.3.0
django-datetime-widget==0.9.3
django-debug-toolbar==1.5
django-filter==1.0.2
django-froala-editor==2.6.4
django-js-asset==0.1.1
django-realtime==1.1
django-tinymce==2.6.0
docker-py==1.10.3
docker-pycreds==0.2.1
docutils==0.13.1
entrypoints==0.2.2
et-xmlfile==1.0.1
face-recognition-models==0.1.3
fastcache==1.0.2
fastzbarlight==0.0.14
Flask==0.10.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.2
Flask-Login==0.2.9
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Stormpath==0.0.1
Flask-Testing==0.6.1
Flask-WTF==0.14.2
gaupol==1.2
gevent==1.2.1
gi==1.2
go-vncdriver==0.4.19
google-api-python-client==1.6.2
greenlet==0.4.12
gym==0.9.2
h5py==2.7.0
HeapDict==1.0.0
hmmlearn==0.2.0
html5lib==0.999999999
httplib2==0.10.3
hyperlink==17.2.1
idna==2.6
imagesize==0.7.1
incremental==17.5.0
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.1.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.9.0
Jinja2==2.7.3
jmespath==0.9.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.8
kombu==3.0.37
lazy-object-proxy==1.2.2
llvmlite==0.16.0
locket==0.2.0
lockfile==0.12.2
lxml==3.7.3
Mako==1.0.6
Markdown==2.6.9
MarkupSafe==0.23
matplotlib==2.0.2
mercurial==4.2
migrate==0.3.8
mistune==0.7.4
mpmath==0.19
multipledispatch==0.4.9
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.2
nose==1.3.7
notebook==4.4.1
numba==0.31.0
numexpr==2.6.2
numpy==1.13.1
numpydoc==0.6.0
oauth2client==4.0.0
odfpy==1.3.5
odo==0.5.0
olefile==0.44
opencv-python==3.2.0.6
openpyxl==2.4.1
pandas==0.19.2
pandocfilters==1.4.1
partd==0.3.7
pathlib2==2.2.0
pathtools==0.1.2
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
plotly==2.0.15
ply==3.10
prompt-toolkit==1.0.13
protobuf==3.3.0
psdash==0.6.2
psutil==5.2.1
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.33
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyface==5.1.0
pyflakes==1.5.0
pyglet==1.2.4
Pygments==2.2.0
PyJWT==1.4.2
pylint==1.6.4
pyOpenSSL==16.2.0
pyparsing==2.1.4
pysolr==3.6.0
pytest==3.0.7
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2014.10
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.2.1
QtPy==1.2.1
redis==2.10.3
reload==0.9
requests==2.18.4
rope-py3k==0.9.4.post1
rsa==3.4.2
ruamel.yaml==0.14.12
s3transfer==0.1.10
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
seaborn==0.7.1
selenium==3.5.0
shadowsocks==2.8.2
simplegeneric==0.8.1
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
sox==1.2.8
Sphinx==1.5.1
spyder==3.1.3
SQLAlchemy==1.1.4
sqlparse==0.2.3
statsmodels==0.8.0
stormpath==1.0.1
sympy==1.0
tables==3.3.0
tablib==0.11.5
tensorflow==1.3.0
tensorflow-tensorboard==0.1.5
terminado==0.6
testpath==0.3
Theano==0.9.0
toolz==0.8.2
tornado==4.4.2
tqdm==4.14.0
traitlets==4.3.2
traits==4.6.0
traitsui==5.1.0
Twisted==17.5.0
txaio==2.8.0
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
vext==0.6.8
vext.gi==0.6.8
vine==1.1.3
vispy==0.5.0.dev0
watchdog==0.8.3
watson-developer-cloud==0.26.1
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.44.0
Werkzeug==0.12.2
widgetsnbextension==2.0.0
wordcloud==1.3.1
wrapt==1.10.8
WTForms==2.1
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
youtube-dl==2017.6.25
zope.interface==4.4.2
